# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NIAT SPI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/filters:
    get:
      operationId: getDashboardFilters
      tags: [dashboard]
      summary: Live campus and section filter options (role-scoped, from BigQuery)
      parameters:
        - name: campus
          in: query
          required: false
          description: When set, sections are limited to this campus
          schema:
            type: string
      responses:
        "200":
          description: Filter options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardFilters"
  /dashboard/students:
    get:
      operationId: getDashboardStudents
      tags: [dashboard]
      summary: Get scoped student list with SPI paths
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: campus
          in: query
          schema:
            type: string
        - name: section
          in: query
          schema:
            type: string
        - name: subject
          in: query
          description: Filter by subject title (subject-wise attendance)
          schema:
            type: string
        - name: attendanceBand
          in: query
          description: "all | below50 | below80 | above80"
          schema:
            type: string
      responses:
        "200":
          description: Student list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DashboardStudent"
...
components:
  schemas:
...
    DashboardFilters:
      type: object
      required: [campuses, sections, updatedAt]
      properties:
        campuses:
          type: array
          items:
            type: string
        sections:
          type: array
          items:
            type: string
        updatedAt:
          type: string
          format: date-time
...
    DashboardStudent:
      type: object
      required: [studentId, studentName, instituteName, sectionName, presentCount, totalCount, attendancePct, spiPath]
      properties:
        studentId:
          type: string
        studentName:
          type: string
        instituteName:
          type: string
        sectionName:
          type: ["string", "null"]
        presentCount:
          type: integer
        totalCount:
          type: integer
        attendancePct:
          type: number
        classroomAvg:
          type: ["number", "null"]
        moduleAvg:
          type: ["number", "null"]
        spiPath:
          type: string
...